requirements.txt
Mastodon.py ~= 1.8.0